ENH: Add answer keys for each fib exercise.
To show path from fib.c -> fib-complete.c, add example files which show the
answers to the exercises for error handling and using the number API.

Each answer key file attempts to minimize the amount of edits needed between the
exercises leading up to the completed fib example.

Intended order of fib examples is:
- fib.c
- fib-error-handling.c
- fib-abstract-api.c
- fib-complete.c

fib-abstract-api-extra.c shows using the fibc function for the values of n which
do not result in a overflow when computing fib(n).